pyproject.toml
[tool.poetry]
name = "hn-clone"
version = "0.1.0"
description = ""
authors = ["Dan Jacob <danjac2018@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.6"
djangorestframework = "^3.11.0"
django-allauth = "^0.41.0"
whitenoise = "^5.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
